app/chatbot/requirements.txt
openai
pinecone
python-dotenv
flask